new commit

+++ routes/web.php
@@ -29,9 +29,11 @@ Route::get('/services', [HomeController::class, 'services']);
 
 Route::get('/blog', [HomeController::class, 'blog']);
 
+// Auth Routes
+
 Route::get('/register', [AuthController::class, 'viewRegister'])->name('register')->middleware('guest');
 
-Route::post('/register', [AuthController::class, 'register'])->name('register.submit')->middleware('auth');
+Route::post('/register', [AuthController::class, 'register'])->name('register.submit')->middleware('guest');
 
 Route::get('/login', [AuthController::class, 'viewLogin'])->name('login')->middleware('guest');
 
@@ -44,16 +46,32 @@ Route::post('/logout', [AuthController::class, 'logout'])->name('logout.submit')
    
 // });
 
+// Admin Routes-User Management
+
 Route::get('/dashboard', [DashboardController::class, 'dashboard'])->name('dashboard')->middleware('auth');
 
-Route::get('/count-users', [DashboardController::class, 'countUsers'])->name('count-users')->middleware('auth');
+Route::get('/count-users', [DashboardController::class, 'countUsers'])->name('count-users')->middleware(['auth', 'permission:getUsers']);
 
-Route::get('/get-users', [DashboardController::class, 'getUsers'])->name('get-users')->middleware('auth');
+Route::get('/get-users', [DashboardController::class, 'getUsers'])->name('get-users')->middleware(['auth', 'permission:getUsers']);
 
-Route::get('/view-users', [DashboardController::class, 'viewUsers'])->name('view-users')->middleware('auth');
+Route::get('/view-users', [DashboardController::class, 'viewUsers'])->name('view-users')->middleware(['auth', 'permission:getUsers']);
 
-Route::put('/users/{id}', [DashboardController::class, 'updateUser'])->name('users.update')->middleware('auth');
+Route::put('/users/{id}', [DashboardController::class, 'updateUser'])->name('users.update')->middleware(['auth', 'permission:editUsers']);
 
-Route::get('/search-users', [DashboardController::class, 'searchUsers'])->name('userSearch')->middleware('auth');
+Route::get('/search-users', [DashboardController::class, 'searchUsers'])->name('userSearch')->middleware(['auth', 'permission:getUsers']);
 
-Route::delete('/delete-user/{id}', [DashboardController::class, 'deleteUser'])->name('users.delete')->middleware('auth');
+Route::delete('/delete-user/{id}', [DashboardController::class, 'deleteUser'])->name('users.delete')->middleware(['auth', 'permission:deleteUsers']);
+
+// Admin Routes-Role Management
+
+Route::post('/create-role', [DashboardController::class, 'createRole'])->name('create-role')->middleware(['auth', 'permission:createRoles']);
+
+Route::put('/edit-role/{id}', [DashboardController::class, 'editRole'])->name('edit-role')->middleware(['auth', 'permission:editRoles']);
+
+Route::delete('/delete-role/{id}', [DashboardController::class, 'deleteRole'])->name('delete-role')->middleware(['auth', 'permission:deleteRoles']);
+
+Route::get('/roles', [DashboardController::class, 'Roles'])->name('roles')->middleware(['auth', 'permission:viewRoles']);
+
+Route::get('/get-roles', [DashboardController::class, 'getRoles'])->name('get-roles')->middleware(['auth', 'permission:viewRoles']);
+
+Route::get('/get-permissions', [DashboardController::class, 'getPermissions'])->name('get-permissions')->middleware(['auth', 'permission:viewRoles']);
